ops: reorder & clean the code
