+++ apps/backend-ops/src/firebaseSetup.ts
@@ -4,99 +4,21 @@
  * This module provides functions to trigger a firestore restore and test user creations.
  */
 import request from 'request';
-import { UserConfig, USERS } from './users.fixture';
-import { differenceBy } from 'lodash';
-import { Auth, loadAdminServices, UserRecord } from './admin';
-
-const sleep = ms => {
-  return new Promise(resolve => {
-    setTimeout(resolve, ms);
-  });
-};
-
-/**
- * @param auth  Firestore Admin Auth object
- * @param uid
- * @param email
- * @param password
- */
-async function createUserIfItDoesntExists(
-  auth: Auth,
-  { uid, email, password }: UserConfig
-): Promise<UserRecord> {
-  try {
-    console.log('trying to get user:', uid, email);
-    // await here to catch the error in the try / catch scope
-    return await auth.getUser(uid);
-  } catch {
-    console.log('creating user:', uid, email);
-    return await auth.createUser({ uid, email, password });
-  }
-}
-
-/**
- * Create all users defined in the users.fixture file
- *
- * @param users The list of users to create if they do not exists.
- * @param auth  Firestore Admin Auth object
- */
-export async function createAllUsers(users: UserConfig[], auth: Auth): Promise<any> {
-  const ps = users.map(user => createUserIfItDoesntExists(auth, user));
-  return Promise.all(ps);
-}
-
-/**
- * Remove all users that are not in the list of expected users.
- *
- * @param expectedUsers
- * @param auth
- */
-export async function removeUnexpectedUsers(expectedUsers: UserConfig[], auth: Auth): Promise<any> {
-  let pageToken;
-
-  do {
-    const result = await auth.listUsers(1000, pageToken);
-
-    const users = result.users;
-    pageToken = result.pageToken;
-
-    console.log('PT:', pageToken);
-    console.log('users:', users.map(x => x.uid));
-
-    // users - expected users => users that we don't want in the database.
-    const usersToRemove = differenceBy(users, expectedUsers, 'uid');
-
-    console.log('usersToRemove:', usersToRemove.map(x => x.uid));
-
-    // Note: this is usually bad practice to await in a loop.
-    // In this VERY SPECIFIC case we just want to remove the user
-    // and wait for some time to avoid exceeding Google's quotas.
-    // This is "good enough", but do not reproduce in frontend / backend code.
-    for (const user of usersToRemove) {
-      console.log('removing user:', user.email, user.uid);
-      await auth.deleteUser(user.uid);
-      await sleep(100);
-    }
-  } while (pageToken);
-
-  return;
-}
-
+import { appUrl } from '@env';
+import { syncUsers } from './users';
+import { upgradeAlgoliaMovies, upgradeAlgoliaOrgs } from './algolia';
+import { migrate } from './migrations';
 
 function getRestoreURL(appURL: string): string {
   return `${appURL}/admin/data/restore`;
 }
 
-export async function syncUsers(expectedUsers: UserConfig[]): Promise<any> {
-  const { auth } = loadAdminServices();
-  await removeUnexpectedUsers(expectedUsers, auth);
-  await createAllUsers(expectedUsers, auth);
+function getBackupURL(appURL: string): string {
+  return `${appURL}/admin/data/backup`;
 }
 
 /**
  * Trigger a firestore database restore operation for the given project
- *
- * @param projectID
  */
 export async function restore(appURL: string) {
   const url = getRestoreURL(appURL);
@@ -114,29 +36,40 @@ export async function restore(appURL: string) {
 }
 
 /**
- * Prepare the project's database & users,
- * run this before testing.
- *
- * TODO: we should be able to disable this operation during certain tests, use an env variable for example.
+ * Trigger a firestore database backup operation for the given project
  */
-export async function prepareFirebase() {
-  const { auth, firebaseConfig } = loadAdminServices();
+export async function backup(appURL: string) {
+  const url = getBackupURL(appURL);
 
-  try {
-    console.info('restoring...');
-    await restore(firebaseConfig.projectId);
-    console.info('done.');
-  } catch (e) {
-    console.error(e);
-  }
+  // promisified request
+  return new Promise((resolve, reject) => {
+    request.post(url, (error, response) => {
+      if (error) {
+        reject(error);
+      } else {
+        resolve(response);
+      }
+    });
+  });
+}
 
-  try {
-    console.info('create all users...');
-    await createAllUsers(USERS, auth);
-    console.info('clearing other users...');
-    await removeUnexpectedUsers(USERS, auth);
-    console.info('done.');
-  } catch (e) {
-    console.error(e);
-  }
+export async function prepareForTesting() {
+  console.info('Syncing users...');
+  await syncUsers();
+  console.info('Users synced!');
+
+  console.info('Restoring backup...');
+  await restore(appUrl);
+  console.info('backup restored!');
+
+  console.info('Preparing Algolia...');
+  await upgradeAlgoliaOrgs();
+  await upgradeAlgoliaMovies();
+  console.info('Algolia ready for testing!');
+
+  console.info('Preparing the database...');
+  await migrate(false); // run the migration, do not trigger a backup before, since we already have it!
+  console.info('database ready for testing...');
+
+  process.exit(0);
 }
