--- apps/backend-ops/src/users.toronto.fixture.ts
@@ -1,97 +1,0 @@
-/**
- * The list of users, ids, passwords to use in the Toronto demos.
- */
-import { UserConfig } from './users.fixture';
-
-export const USERS: UserConfig[] = [
-  {
-    uid: '8sIkEsgILVNIk6gET7SyuphCrGp1',
-    email: 'cytest@blockframes.com',
-    password: 'azerty'
-  },
-  {
-    uid: 'uJY5HNFXj8PUq8CcQvqni1Ywm0s1',
-    email: 'cytest.thejokers@blockframes.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'mykt2DfYNFNvLeslFqY8R2GRnhm1',
-    email: 'cytest.cjentertainment@blockframes.com',
-    password: 'blockframes'
-  },
-  {
-    uid: '1H0F5NJqfkMHRwLIqAnYbUECiO22',
-    email: 'forfun@vincent.fr',
-    password: 'blockframes'
-  },
-  {
-    uid: '1cDhS7rdLeZARDuRDixCxtPBUe42',
-    email: 'hello@clockworks.com',
-    password: 'blockframes'
-  },
-  {
-    uid: '3MjmItCZi2Nghuvks91frDP8hQ03',
-    email: 'hello@syncopatedfilms.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'CART4GbCHPTrpWgAbZtsmUyfwly1',
-    email: 'dazdad@gmail.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'CVE2GVi44SgwtKA0YUaOe5nfamy2',
-    email: 'hello@netflix.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'EDIWkg04UeTBgAEyVA88RtTJmUr1',
-    email: 'ddada@test.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'LKKobibLhCXxlPuDJwrxqDalSjg2',
-    email: 'hello@barunsonea.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'UCcHMvd398TTc3wsJ3KKewdy3t63',
-    email: 'hello@empire.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'UsYWmseMoOQVN2hWrJ8oxlaQm0C2',
-    email: 'hello@wearefabula.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'bPpo9XhbhpZ1Ic1RGkWKYzP7WBJ2',
-    email: 'hello@logicalpictures.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'fqK319OtikZsW4TbV3tnVXVC6hs1',
-    email: 'hello@charades.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'jJKxS7Ix2VP1fYeTCaLVSR3sk4F3',
-    email: 'hello@volgafilms.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'qMywJAW4VtY4uG5iv44AAxCLK9D3',
-    email: 'hello@thejokers.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'XfE8uBoeFRToxrPipXOHPgZJG9U2',
-    email: 'hello@cascade8.com',
-    password: 'blockframes'
-  },
-  {
-    uid: 'rjcV3BfLGlPalqUrbRX91zWqwqB2',
-    email: 'cjentertainment@blockframes.com',
-    password: 'blockframes'
-  }
-];
